Fixed a bug in the last rev. Player can run now >P Added npcs in Shilon, Paros, Dekadun, Searost, Dimitron, Great White Horn, Dekardi and Laglamia
git-svn-id: http://reunion.googlecode.com/svn/trunk@306 d5b2e400-7aba-11de-9499-4f8beaa274e3

+++ jreunion/src/com/googlecode/reunion/jreunion/server/PacketFactory.java
@@ -333,10 +333,10 @@ public class PacketFactory {
 		case STATUS:
 			if(args.length>1){
 				int id = (Integer)args[0];
-				float arg1 = (Float)args[1];
+				float arg1 = (Long)args[1];
 				float arg2 = 0;
 				if(args.length > 2){
-					arg2 = (Float)args[2];
+					arg2 = (Long)args[2];
 				}
 				return "status " + id + " " + arg1 + " " + arg2;
 			}
